Conform legal_markdown.rb to Sublime version
As Legal-Markdown-Sublime was processing files successfully, I
cross-checked that repository's version of legal_markdown.rb against the
version that was not working for me in the Legal-Markdown repository.

That was progress, but not the end of the story. At the command prompt,
`legal2md` was still throwing `NoMethodError` exceptions on the
`@headers` in `leaders.rb`. This appears to be because my input file
lacked a complete set of YAML front-matter. Once I included at least
placeholders for the three `Properties` headers (`no-indent`,
`no-reset`, and `level-style`) I was successful in running `legal2md
[input file] [output file]` and `legal2md -b [input file] [output file]`
from the Windows 7 Command Prompt without it returning the previous
error, which was: `Sorry, I could not read the file ["--headers",
"test.lmd"]: no implicit conversion of Array into String.`